Files
This is a simple Quiz App with integrated API from OpenTDB. It has 10 questions and at the end your final score is showed based on your correct answers. The simple and minimalistic yet powerful design, makes it easy to use on every device. Hope you like it and thanks for visiting this repository!